refactor(#293): 버전 명시, 리소스 복수형으로 개선
