+++ src/main/java/com/sideproject/cafe_cok/member/presentation/MemberController.java
@@ -5,6 +5,9 @@ import com.sideproject.cafe_cok.auth.presentation.AuthenticationPrincipal;
 import com.sideproject.cafe_cok.member.application.MemberService;
 import com.sideproject.cafe_cok.member.dto.request.MemberFeedbackRequest;
 import com.sideproject.cafe_cok.member.dto.response.MemberResponse;
+import com.sideproject.cafe_cok.plan.domain.enums.PlanSortBy;
+import com.sideproject.cafe_cok.plan.domain.enums.PlanStatus;
+import com.sideproject.cafe_cok.plan.dto.response.PlanAllResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +21,7 @@ import java.io.IOException;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/api/member")
+@RequestMapping("/api/v1/members")
 @Tag(name = "Member", description = "회원 관련 API")
 public class MemberController {
 
@@ -41,7 +44,7 @@ public class MemberController {
         return ResponseEntity.ok(response);
     }
 
-    @PostMapping("/feedback")
+    @PostMapping("/feedbacks")
     @Operation(summary = "회원 개선의견 저장")
     public ResponseEntity<Void> saveFeedback(@AuthenticationPrincipal LoginMember loginMember,
                                              @RequestBody MemberFeedbackRequest request) {
@@ -50,5 +53,13 @@ public class MemberController {
         return ResponseEntity.noContent().build();
     }
 
+    @GetMapping("/plans")
+    @Operation(summary = "회원의 모든 계획 조회")
+    public ResponseEntity<PlanAllResponse> findPlans(@AuthenticationPrincipal LoginMember loginMember,
+                                                     @RequestParam(defaultValue = "RECENT") PlanSortBy sortBy,
+                                                     @RequestParam(defaultValue = "SAVED") PlanStatus status) {
 
+        PlanAllResponse response = memberService.findPlans(loginMember, sortBy, status);
+        return ResponseEntity.ok(response);
+    }
 }
