+++ src/main/java/com/sideproject/cafe_cok/plan/application/PlanService.java
@@ -11,24 +11,18 @@ import com.sideproject.cafe_cok.keword.domain.enums.Category;
 import com.sideproject.cafe_cok.keword.domain.repository.KeywordRepository;
 import com.sideproject.cafe_cok.keword.dto.CategoryKeywordsDto;
 import com.sideproject.cafe_cok.member.domain.repository.MemberRepository;
-import com.sideproject.cafe_cok.plan.domain.condition.PlanSearchCondition;
 import com.sideproject.cafe_cok.plan.dto.response.PlanResponse;
-import com.sideproject.cafe_cok.plan.dto.response.PlanAllResponse;
-import com.sideproject.cafe_cok.plan.dto.response.PlanPageResponse;
 import com.sideproject.cafe_cok.plan.domain.Plan;
 import com.sideproject.cafe_cok.plan.domain.PlanCafe;
 import com.sideproject.cafe_cok.plan.domain.PlanKeyword;
 import com.sideproject.cafe_cok.plan.domain.enums.MatchType;
-import com.sideproject.cafe_cok.plan.domain.enums.PlanSortBy;
 import com.sideproject.cafe_cok.plan.domain.enums.PlanStatus;
 import com.sideproject.cafe_cok.plan.domain.repository.PlanCafeRepository;
 import com.sideproject.cafe_cok.plan.domain.repository.PlanKeywordRepository;
 import com.sideproject.cafe_cok.plan.domain.repository.PlanRepository;
-import com.sideproject.cafe_cok.plan.dto.PlanKeywordDto;
 import com.sideproject.cafe_cok.plan.dto.request.PlanSaveRequest;
 import com.sideproject.cafe_cok.plan.dto.response.SavePlanResponse;
 import com.sideproject.cafe_cok.plan.dto.response.PlanIdResponse;
-import com.sideproject.cafe_cok.plan.exception.NoSuchPlanSortException;
 import com.sideproject.cafe_cok.utils.Constants;
 import com.sideproject.cafe_cok.utils.ListUtils;
 import com.sideproject.cafe_cok.utils.tmap.client.TmapClient;
@@ -37,8 +31,6 @@ import com.sideproject.cafe_cok.member.domain.Member;
 import com.sideproject.cafe_cok.cafe.domain.Cafe;
 import com.sideproject.cafe_cok.plan.exception.NoSuchPlanKeywordException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -47,8 +39,6 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 import static com.sideproject.cafe_cok.utils.FormatConverter.convertLocalDateLocalTimeToString;
-import static org.springframework.data.domain.Sort.by;
-
 
 @Service
 @RequiredArgsConstructor
@@ -65,9 +55,6 @@ public class PlanService {
     private final PlanKeywordRepository planKeywordRepository;
     private final BookmarkRepository bookmarkRepository;
     private final ImageRepository imageRepository;
-
-    private final Integer FIRST_PAGE_NUMBER = 0;
-    private final Integer MAX_PAGE_SIZE = Integer.MAX_VALUE;
 
     @Transactional
     public SavePlanResponse doPlan(final PlanSaveRequest request,
@@ -171,34 +158,6 @@ public class PlanService {
         return new PlanIdResponse(findPlan.getId());
     }
 
-    public PlanPageResponse getPlans(final LoginMember loginMember,
-                                     final PlanSortBy planSortBy,
-                                     final PlanStatus status,
-                                     final Integer page,
-                                     final Integer size) {
-
-        PlanSearchCondition planSearchCondition
-                = new PlanSearchCondition(loginMember.getId(), Category.PURPOSE, planSortBy, status);
-        Sort sort = getSort(planSortBy);
-        Pageable pageable = PageRequest.of(page-1, size, sort);
-        List<PlanKeywordDto> plans = planRepository.findPlanKeywordDtoList(planSearchCondition, pageable);
-
-        return new PlanPageResponse(page, plans);
-    }
-
-    public PlanAllResponse getPlansAll(final LoginMember loginMember,
-                                       final PlanSortBy planSortBy,
-                                       final PlanStatus status) {
-
-        PlanSearchCondition planSearchCondition
-                = new PlanSearchCondition(loginMember.getId(), Category.PURPOSE, planSortBy, status);
-        Sort sort = getSort(planSortBy);
-        Pageable pageable = PageRequest.of(FIRST_PAGE_NUMBER, MAX_PAGE_SIZE, sort);
-        List<PlanKeywordDto> plans = planRepository.findPlanKeywordDtoList(planSearchCondition, pageable);
-
-        return new PlanAllResponse(plans);
-    }
-
     public PlanResponse findPlan(final LoginMember loginMember,
                                  final Long planId) {
 
@@ -229,17 +188,6 @@ public class PlanService {
                     return new CafeDto(cafe, findImageUrl, findBookmarkIdDtoList);
                 }).collect(Collectors.toList());
         return cafeDtoList;
-    }
-
-    private Sort getSort(final PlanSortBy planSortBy) {
-        Sort sort;
-        switch (planSortBy){
-            case RECENT -> sort = by(Sort.Direction.DESC, PlanSortBy.RECENT.getValue());
-            case UPCOMING -> sort = by(Sort.Direction.ASC, PlanSortBy.UPCOMING.getValue(), "visitStartTime", "id");
-            default -> throw new NoSuchPlanSortException();
-        }
-
-        return sort;
     }
 
     private SavePlanResponse createMisMatchPlan(final PlanSaveRequest request,
