+++ src/main/java/com/sideproject/cafe_cok/plan/presentation/PlanController.java
@@ -7,11 +7,8 @@ import com.sideproject.cafe_cok.auth.exception.InvalidTokenException;
 import com.sideproject.cafe_cok.auth.presentation.AuthenticationPrincipal;
 import com.sideproject.cafe_cok.auth.presentation.AuthorizationExtractor;
 import com.sideproject.cafe_cok.plan.dto.response.PlanResponse;
-import com.sideproject.cafe_cok.plan.dto.response.PlanAllResponse;
-import com.sideproject.cafe_cok.plan.dto.response.PlanPageResponse;
 import com.sideproject.cafe_cok.member.exception.NoSuchMemberException;
 import com.sideproject.cafe_cok.plan.application.PlanService;
-import com.sideproject.cafe_cok.plan.domain.enums.PlanSortBy;
 import com.sideproject.cafe_cok.plan.domain.enums.PlanStatus;
 import com.sideproject.cafe_cok.plan.dto.request.PlanSaveRequest;
 import com.sideproject.cafe_cok.plan.dto.response.SavePlanResponse;
@@ -68,46 +65,6 @@ public class PlanController {
                                                      @RequestParam PlanStatus status){
 
         PlanIdResponse response = planService.delete(status, planId);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/shared")
-    @Operation(summary = "공유된 계획 조회(페이징)")
-    public ResponseEntity<PlanPageResponse> sharedPlans(@AuthenticationPrincipal LoginMember loginMember,
-                                                        @RequestParam(defaultValue = "RECENT") PlanSortBy sortBy,
-                                                        @RequestParam(defaultValue = "1") Integer page,
-                                                        @RequestParam(defaultValue = "10") Integer size) {
-
-        PlanPageResponse response = planService.getPlans(loginMember, sortBy, PlanStatus.SHARED, page, size);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/saved")
-    @Operation(summary = "저장된 계획 조회(페이징)")
-    public ResponseEntity<PlanPageResponse> savedPlans(@AuthenticationPrincipal LoginMember loginMember,
-                                                       @RequestParam(defaultValue = "RECENT") PlanSortBy sortBy,
-                                                       @RequestParam(defaultValue = "1") Integer page,
-                                                       @RequestParam(defaultValue = "10") Integer size) {
-
-        PlanPageResponse response = planService.getPlans(loginMember, sortBy, PlanStatus.SAVED, page, size);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/saved/all")
-    @Operation(summary = "저장된 계획 조회(전체)")
-    public ResponseEntity<PlanAllResponse> savedPlansAll(@AuthenticationPrincipal LoginMember loginMember,
-                                                         @RequestParam(defaultValue = "RECENT") PlanSortBy sortBy) {
-
-        PlanAllResponse response = planService.getPlansAll(loginMember, sortBy, PlanStatus.SAVED);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/shared/all")
-    @Operation(summary = "공유된 계획 조회(전체)")
-    public ResponseEntity<PlanAllResponse> sharedPlansAll(@AuthenticationPrincipal LoginMember loginMember,
-                                                          @RequestParam(defaultValue = "RECENT") PlanSortBy sortBy) {
-
-        PlanAllResponse response = planService.getPlansAll(loginMember, sortBy, PlanStatus.SHARED);
         return ResponseEntity.ok(response);
     }
